+++ src/HVDCWISE_TEA.jl
@@ -49,11 +49,13 @@ include("form/dcp.jl")
 include("io/build_outputs.jl")
 include("io/build_raw_inputs.jl")
 include("io/build_simulation_inputs.jl")
+include("io/build_user_results.jl")
 include("io/load_case.jl")
 include("io/export.jl")
 include("io/multinetwork.jl")
 include("io/parse.jl")
 include("io/scenario.jl")
+include("io/run_study.jl")
 
 include("parallel/base.jl")